+++ src/main.ts
@@ -1,6 +1,6 @@
-import { OMSLogger } from '@ffdc/logger';
-import { LoggingInterceptor } from '@ffdc/logger/interceptors/common-http-service.interceptor';
-import { setupSession, TokenGuard } from '@ffdc/nestjs-oidc';
+import { OMSLogger } from '@finastra/logger';
+import { HttpLoggingInterceptor } from '@finastra/logger/interceptors/common-http.interceptor';
+import { setupSession, TokenGuard } from '@finastra/nestjs-oidc';
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 
@@ -10,7 +10,7 @@ async function bootstrap() {
     logger: omsLogger,
   });
   app.useLogger(omsLogger);
-  app.useGlobalInterceptors(new LoggingInterceptor(omsLogger));
+  app.useGlobalInterceptors(new HttpLoggingInterceptor());
 
   app.useGlobalGuards(app.get(TokenGuard));